Reorganise preferences
