+++ src/components/preferences/base/preferenceComponent.jsx
@@ -1,10 +1,11 @@
 import { Action, ActionPanel, List } from "@raycast/api";
 
-export const PreferenceComponent = ({ title, target, accessories = [] }) => {
+export const PreferenceComponent = ({ title, subtitle, target, accessories = [] }) => {
   return (
     <List.Item
       title={title}
       key={title}
+      subtitle={subtitle}
       accessories={accessories}
       actions={
         <ActionPanel>
