+++ src/components/preferences/options/aiChatOptions.jsx
@@ -30,33 +30,9 @@ const aiChatOptions = {
     required: false,
     type: "checkbox",
   },
-  autoCheckForUpdates: {
-    title: "Automatically Check for Updates",
-    description: "Automatically check for updates from the official GitHub source.",
-    required: false,
-    type: "checkbox",
-  },
-  persistentStorage: {
-    title: "Enable Persistent Storage",
-    description: "Enable more persistent storage of the extension's data. Do not use if you have sensitive data.",
-    required: false,
-    type: "checkbox",
-  },
-  useCursorIcon: {
-    title: "Enable Cursor Icon",
-    description: "Show a cursor icon when the response is loading.",
-    required: false,
-    type: "checkbox",
-  },
   codeInterpreter: {
     title: "Enable Code Interpreter (BETA)",
     description: "Allows GPT to interpret and run Python code locally, based on your queries. Use at your own risk!",
-    required: false,
-    type: "checkbox",
-  },
-  devMode: {
-    title: "Enable Developer Mode",
-    description: "Enable developer mode for more extensive logging. Do not enable this.",
     required: false,
     type: "checkbox",
   },
